ColumnNumber and ColumnNumberDelta: Migrate to GhostType.
Some things are a bit sub-optimal, should be improved in follow up changes
(removing custom methods from ColumnNumber and ColumnNumberDelta), but reducing
dependencies on GHOST_TYPE macros is good.

+++ src/widget_list.cc
@@ -218,7 +218,7 @@ LineWithCursor::Generator::Vector WidgetListVertical::CreateOutput(
 
   ColumnNumberDelta base_columns = options.size.column / children_.size();
   ColumnNumberDelta columns_left =
-      options.size.column - base_columns * children_.size();
+      options.size.column - base_columns * static_cast<int>(children_.size());
   CHECK_LT(columns_left, ColumnNumberDelta(children_.size()));
   for (auto& column : columns_vector.columns) {
     column.width = base_columns;
